Rename endpoint for oppilaitoksen opiskelijat
