+++ src/main/scala/ScalatraBootstrap.scala
@@ -7,30 +7,17 @@ import fi.vm.sade.hakurekisteri.integration.henkilo.PersonOidsWithAliases
 import fi.vm.sade.hakurekisteri.web.HakuJaValintarekisteriStack
 import fi.vm.sade.hakurekisteri.web.arvosana.{ArvosanaResource, EmptyLisatiedotResource}
 import fi.vm.sade.hakurekisteri.web.ensikertalainen.EnsikertalainenResource
-import fi.vm.sade.hakurekisteri.web.hakija.{
-  HakijaResource,
-  HakijaResourceV2,
-  HakijaResourceV3,
-  HakijaResourceV4,
-  HakijaResourceV5
-}
+import fi.vm.sade.hakurekisteri.web.hakija.{HakijaResource, HakijaResourceV2, HakijaResourceV3, HakijaResourceV4, HakijaResourceV5}
 import fi.vm.sade.hakurekisteri.web.haku.HakuResource
 import fi.vm.sade.hakurekisteri.web.integration.virta.{VirtaResource, VirtaSuoritusResource}
 import fi.vm.sade.hakurekisteri.web.integration.ytl.YtlResource
 import fi.vm.sade.hakurekisteri.web.jonotus.{AsiakirjaResource, SiirtotiedostojonoResource}
-import fi.vm.sade.hakurekisteri.web.kkhakija.{
-  KkHakijaResource,
-  KkHakijaResourceV2,
-  KkHakijaResourceV3,
-  KkHakijaResourceV4
-}
+import fi.vm.sade.hakurekisteri.web.kkhakija.{KkHakijaResource, KkHakijaResourceV2, KkHakijaResourceV3, KkHakijaResourceV4}
 import fi.vm.sade.hakurekisteri.web.koski.KoskiImporterResource
-import fi.vm.sade.hakurekisteri.web.opiskelija.{
-  OpiskelijaResource,
-  OppilaitoksenOpiskelijatResource
-}
+import fi.vm.sade.hakurekisteri.web.opiskelija.OpiskelijaResource
 import fi.vm.sade.hakurekisteri.web.opiskeluoikeus.OpiskeluoikeusResource
 import fi.vm.sade.hakurekisteri.web.oppija.OppijaResource
+import fi.vm.sade.hakurekisteri.web.oppilaitos.OppilaitoksenOpiskelijatResource
 import fi.vm.sade.hakurekisteri.web.permission.PermissionResource
 import fi.vm.sade.hakurekisteri.web.proxies._
 import fi.vm.sade.hakurekisteri.web.rekisteritiedot.RekisteritiedotResource
@@ -171,8 +158,8 @@ class ScalatraBootstrap extends LifeCycle {
       authorizedRegisters.opiskelijaRekisteri
     ),
     (
-      "/rest/v1/opiskelijat/oppilaitos",
-      "rest/v1/opiskelijat/oppilaitos"
+      "/rest/v1/oppilaitos",
+      "rest/v1/oppilaitos"
     ) -> new OppilaitoksenOpiskelijatResource(
       authorizedRegisters.opiskelijaRekisteri
     ),
